+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Reconnect to inverter after 15 minutes of not receiving any data (#223)
 * Fix max/min cell temperature/voltage decoding as reported from BMS (#227)
+* Add more HA entities: max/min cell temp/voltage, more charge powers (#228)
 
 
 # 0.13.0 - 27th October 2023