+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix max/min cell temperature/voltage decoding as reported from BMS (#227)
 * Add more HA entities: max/min cell temp/voltage, more charge powers (#228)
 * Add ReadInput4 with EG4 18k generator data (#239, @pmccut)
+* Add ReadInput4 keys to HA discovery (#240, @jgulick48)
 
 
 # 0.13.0 - 27th October 2023