+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add more HA entities: max/min cell temp/voltage, more charge powers (#228)
 * Add ReadInput4 with EG4 18k generator data (#239, @pmccut)
 * Add ReadInput4 keys to HA discovery (#240, @jgulick48)
+* Fix min_chg_curr/max_chg_curr decoding in ReadInputAll packet (#242, @presto8)
 
 
 # 0.13.0 - 27th October 2023